Make common ancestor for RoutePresenter and AutofillRoutePresenter.
